requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
tenacity>=8.2.3
playwright>=1.40.0
setuptools<81
groq>=0.11.0
